Added widgets
I have added some widgets to the dashboard. There are 4 of them. One for orders, one for average order cost, one for total number of products, and one for the total inventory. These are meant to be general and not specific data, since this is the overview board. I have also done a lot of work on responsiveness. There were so many problems getting this to work on different screen sizes. I have made it so it looks good on iPhone X, iPad, iPad Pro, and regular desktop monitors.