Chart Changes
I have done a few things to the orders chart of the main overview. I have changed the colors of it to fit the brand colors of yellow and gray. I also moved the JavaScript for the chart into its own seperate file that is loaded in to clean up the index.html file. There was also a minor problem with moving the index.html out of the src folder. The avatar image wasn't loading in, but that was a quick path fix.